+++ src/Model/Billing.php
@@ -6,5 +6,32 @@ use Easir\SDK\Model;
 
 class Billing extends Model
 {
-    public $invoice_email, $address_1, $address_2, $zip_code, $city, $state, $country;
+    /**
+     * @var string
+     */
+    public $invoice_email;
+    /**
+     * @var string
+     */
+    public $address_1;
+    /**
+     * @var string
+     */
+    public $address_2;
+    /**
+     * @var string
+     */
+    public $zip_code;
+    /**
+     * @var string
+     */
+    public $city;
+    /**
+     * @var string
+     */
+    public $state;
+    /**
+     * @var Country
+     */
+    public $country;
 }