+++ src/Model/Delay.php
@@ -6,5 +6,12 @@ use Easir\SDK\Model;
 
 class Delay extends Model
 {
-    public $days, $offset_from_midnight;
+    /**
+     * @var int
+     */
+    public $days;
+    /**
+     * @var int
+     */
+    public $offset_from_midnight;
 }