+++ src/Model/RelationsPath.php
@@ -6,7 +6,36 @@ use Easir\SDK\Response;
 
 class RelationsPath extends Response
 {
-    protected $collections = ['activity_types' => 'activity_type'];
-
-    public $id, $name, $is_sales_type, $activity_types, $team_id, $group_id, $created_at, $updated_at;
+    /**
+     * @var string
+     */
+    public $id;
+    /**
+     * @var string
+     */
+    public $name;
+    /**
+     * @var bool
+     */
+    public $is_sales_type;
+    /**
+     * @var ActivityTypes
+     */
+    public $activity_types;
+    /**
+     * @var int|null
+     */
+    public $team_id;
+    /**
+     * @var int|null
+     */
+    public $group_id;
+    /**
+     * @var string
+     */
+    public $created_at;
+    /**
+     * @var string
+     */
+    public $updated_at;
 }