+++ src/Model/Template.php
@@ -2,10 +2,60 @@
 
 namespace Easir\SDK\Model;
 
-use Easir\SDK\Model;
+use Easir\SDK\Response;
 
-class Template extends Model
+class Template extends Response
 {
-    public $id, $type, $company_id, $title, $user_id, $team_id, $group_id, $campaign, $margin_top, $margin_left,
-        $margin_right, $margin_bottom;
+    /**
+     * @var int
+     */
+    public $id;
+    /**
+     * @var string
+     */
+    public $type;
+    /**
+     * @var int
+     */
+    public $company_id;
+    /**
+     * @var string
+     */
+    public $title;
+    /**
+     * @var int
+     */
+    public $user_id;
+    /**
+     * @var int|null
+     */
+    public $team_id;
+    /**
+     * @var int|null
+     */
+    public $group_id;
+    /**
+     * @var bool
+     */
+    public $campaign;
+    /**
+     * @var int|null
+     */
+    public $margin_top;
+    /**
+     * @var int|null
+     */
+    public $margin_left;
+    /**
+     * @var int|null
+     */
+    public $margin_right;
+    /**
+     * @var int|null
+     */
+    public $margin_bottom;
+    /**
+     * @var string
+     */
+    public $content;
 }